Revert "reup"
This reverts commit 40ce9d5afde993bb50b133a4f0f4e20f05241945.

+++ database/migrations/2025_02_08_161356_add_user_catalogue_id_to_users.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             
-            $table->bigInteger('user_catalogue_id')->default(2);
+            $table->integer('user_catalogue_id')->default(2);
         });
     }
